Fix VSM tests and add utils directory
