+++ tests/queries.json
@@ -3,44 +3,33 @@
          {
             "query": "footballer France Forward",
             "result": [
-                "karim_benzema.txt",
-                "antoine_griezmann.txt",
-                "kylian_mbappe.txt"
+                "Karim_Benzema",
+                "Antoine_Griezmann",
+                "Kylian_Mbappe"
             ]
         },
         {
             "query": "footballer Barcelona forward",
             "result": [
-                "lionel_messi.txt",
-                "neymar.txt",
-                "antoine_griezmann.txt"
-            ]
-        },
-        {
-            "query": "footballer born march",
-            "result": [
-                "antoine_griezmann.txt",
-                "paul_pogba.txt",
-                "manuel_neuer.txt",
-                "hakim_ziyech.txt",
-                "timo_werner.txt",
-                "carlo_pinsoglio.txt"
+                "Lionel_Messi",
+                "Neymar",
+                "Antoine_Griezmann"
             ]
         },
         {
             "query": "footballer Bayern",
             "result": [
-                "manuel_neuer.txt",
-                "robert_lewandowski.txt",
-                "david_alaba.txt",
-                "toni_kroos.txt"
+                "Manuel_Neuer",
+                "Robert_Lewandowski",
+                "David_Alaba",
+                "Toni_Kroos"
             ]
         },
         {
             "query": "footballer defensive midfielder",
             "result": [
-                "casemiro.txt",
-                "trevoh_chalobah.txt"
+                "Casemiro",
+                "Trevoh_Chalobah"
             ]
         }
     ]
